Created database class Database based on sqlite moved endpoints to use database class moved block chain to use database class
This deprecates the Berkeley DB

+++ include/coin/BigNum.h
@@ -214,12 +214,12 @@ public:
     {
         unsigned int nSize = BN_bn2mpi(this, NULL);
         if (nSize < 4)
-            return 0;
+            return uint256(0);
         std::vector<unsigned char> vch(nSize);
         BN_bn2mpi(this, &vch[0]);
         if (vch.size() > 4)
             vch[4] &= 0x7f;
-        uint256 n = 0;
+        uint256 n = uint256(0);
         for (int i = 0, j = vch.size()-1; i < sizeof(n) && j >= 4; i++, j--)
             ((unsigned char*)&n)[i] = vch[j];
         return n;
